LPS-127220 Fix tests
