+++ modules/dxp/apps/analytics/analytics-reports-test/src/testIntegration/java/com/liferay/analytics/reports/test/util/MockContextUtil.java
@@ -17,9 +17,7 @@ package com.liferay.analytics.reports.test.util;
 import com.liferay.analytics.reports.info.item.AnalyticsReportsInfoItem;
 import com.liferay.analytics.reports.test.MockObject;
 import com.liferay.analytics.reports.test.analytics.reports.info.item.MockAnalyticsReportsInfoItem;
-import com.liferay.analytics.reports.test.info.item.provider.MockInfoItemFieldValuesProvider;
 import com.liferay.analytics.reports.test.layout.display.page.MockLayoutDisplayPageProvider;
-import com.liferay.info.item.provider.InfoItemFieldValuesProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageProvider;
 import com.liferay.petra.function.UnsafeRunnable;
 import com.liferay.portal.kernel.model.ClassName;
@@ -52,8 +50,6 @@ public class MockContextUtil {
 
 		ServiceRegistration<AnalyticsReportsInfoItem<MockObject>>
 			analyticsReportsInfoItemServiceRegistration = null;
-		ServiceRegistration<InfoItemFieldValuesProvider<MockObject>>
-			infoItemFieldValuesProviderServiceRegistration = null;
 		ServiceRegistration<LayoutDisplayPageProvider<MockObject>>
 			layoutDisplayPageProviderServiceRegistration = null;
 
@@ -63,12 +59,6 @@ public class MockContextUtil {
 					(Class<AnalyticsReportsInfoItem<MockObject>>)
 						(Class<?>)AnalyticsReportsInfoItem.class,
 					mockContext.getAnalyticsReportsInfoItem(),
-					new HashMapDictionary<>());
-			infoItemFieldValuesProviderServiceRegistration =
-				bundleContext.registerService(
-					(Class<InfoItemFieldValuesProvider<MockObject>>)
-						(Class<?>)InfoItemFieldValuesProvider.class,
-					mockContext.getInfoItemFieldValuesProvider(),
 					new HashMapDictionary<>());
 			layoutDisplayPageProviderServiceRegistration =
 				bundleContext.registerService(
@@ -82,10 +72,6 @@ public class MockContextUtil {
 		finally {
 			if (analyticsReportsInfoItemServiceRegistration != null) {
 				analyticsReportsInfoItemServiceRegistration.unregister();
-			}
-
-			if (infoItemFieldValuesProviderServiceRegistration != null) {
-				infoItemFieldValuesProviderServiceRegistration.unregister();
 			}
 
 			if (layoutDisplayPageProviderServiceRegistration != null) {
@@ -114,12 +100,6 @@ public class MockContextUtil {
 			return _classNameLocalService;
 		}
 
-		public InfoItemFieldValuesProvider<MockObject>
-			getInfoItemFieldValuesProvider() {
-
-			return _infoItemFieldValuesProvider;
-		}
-
 		public LayoutDisplayPageProvider<MockObject>
 			getLayoutDisplayPageProvider() {
 
@@ -143,16 +123,7 @@ public class MockContextUtil {
 			public MockContext build() {
 				return new MockContext(
 					_analyticsReportsInfoItem, _classNameLocalService,
-					_infoItemFieldValuesProvider, _layoutDisplayPageProvider);
-			}
-
-			public Builder infoItemFieldValuesProvider(
-				InfoItemFieldValuesProvider<MockObject>
-					infoItemFieldValuesProvider) {
-
-				_infoItemFieldValuesProvider = infoItemFieldValuesProvider;
-
-				return this;
+					_layoutDisplayPageProvider);
 			}
 
 			public Builder layoutDisplayPageProvider(
@@ -167,8 +138,6 @@ public class MockContextUtil {
 			private AnalyticsReportsInfoItem<MockObject>
 				_analyticsReportsInfoItem;
 			private final ClassNameLocalService _classNameLocalService;
-			private InfoItemFieldValuesProvider<MockObject>
-				_infoItemFieldValuesProvider;
 			private LayoutDisplayPageProvider<MockObject>
 				_layoutDisplayPageProvider;
 
@@ -177,7 +146,6 @@ public class MockContextUtil {
 		private MockContext(
 			AnalyticsReportsInfoItem<MockObject> analyticsReportsInfoItem,
 			ClassNameLocalService classNameLocalService,
-			InfoItemFieldValuesProvider<MockObject> infoItemFieldValuesProvider,
 			LayoutDisplayPageProvider<MockObject> layoutDisplayPageProvider) {
 
 			if (analyticsReportsInfoItem == null) {
@@ -190,15 +158,6 @@ public class MockContextUtil {
 			}
 
 			_classNameLocalService = classNameLocalService;
-
-			if (infoItemFieldValuesProvider == null) {
-				_infoItemFieldValuesProvider =
-					MockInfoItemFieldValuesProvider.builder(
-					).build();
-			}
-			else {
-				_infoItemFieldValuesProvider = infoItemFieldValuesProvider;
-			}
 
 			if (layoutDisplayPageProvider == null) {
 				_layoutDisplayPageProvider =
@@ -214,8 +173,6 @@ public class MockContextUtil {
 		private final AnalyticsReportsInfoItem<MockObject>
 			_analyticsReportsInfoItem;
 		private final ClassNameLocalService _classNameLocalService;
-		private final InfoItemFieldValuesProvider<MockObject>
-			_infoItemFieldValuesProvider;
 		private final LayoutDisplayPageProvider<MockObject>
 			_layoutDisplayPageProvider;
